+++ lib/ex_admin/repo.ex
@@ -9,6 +9,9 @@ defmodule ExAdmin.Repo do
 
   def repo, do: Application.get_env(:ex_admin, :repo)
 
+  defp stringify_key(key) when is_atom(key), do: Atom.to_string(key)
+  defp stringify_key(key) when is_binary(key), do: key
+
   def changeset(fun, resource, nil), do: changeset(fun, resource, %{})
   def changeset(fun, resource, params) do
     %ExAdmin.Changeset{}
@@ -20,7 +23,7 @@ defmodule ExAdmin.Repo do
   def changeset(%Changeset{} = changeset, fun, resource, nil),
     do: changeset(changeset, fun, resource, %{})
   def changeset(%Changeset{} = changeset, fun, resource, params) do
-    cs = fun.(resource, params)
+    cs = fun.(resource, param_stringify_keys(params))
     Changeset.update(changeset, valid?: cs.valid?, changeset: cs, errors: cs.errors)
   end
 
@@ -351,5 +354,12 @@ res
     |> Enum.sort(&(elem(&1, 0) < elem(&2, 0)))
     res
   end
+
+  def param_stringify_keys(params) when is_map(params) do
+    Map.to_list(params)
+    |> Enum.map(fn {key, value} -> {stringify_key(key), param_stringify_keys(value)} end)
+    |> Enum.into(%{})
+  end
+  def param_stringify_keys(params), do: params
 
 end